Fix starting items, goal location, and missing fish
Changed starting items to give only one character and one location rather than all of them.
Added new "Victory" location and "You Win!" item to allow the name and description for the actual goal location to be seen.  The victory location is now to be checked whenever the "You Win!" item is received, which is always behind the "1st place in Regional Intermediate Tournament" location.
Added name of missing fish (Oscar).